add game speed

+++ app/src/main/java/ge/gogichaishvili/lotto/main/enums/GameSpeedEnum.kt
@@ -1,13 +1,13 @@
 package ge.gogichaishvili.lotto.main.enums
 
-enum class GameSpeedEnum (val value: Int) {
+enum class GameSpeedEnum (val value: Long) {
 
     LOW(3000),
     MEDIUM(2000),
     HIGH(1500);
 
     companion object {
-        fun getEnumByCode(code: Int): GameSpeedEnum {
+        fun getEnumByCode(code: Long): GameSpeedEnum {
             return when (code) {
                 LOW.value -> LOW
                 MEDIUM.value -> MEDIUM
